blockchain: Move UtxoEntry to a separate file.
This moves the UtxoEntry type and its methods to a separate file in
preparation for adding a utxo cache since both utxo views and the utxo
cache will hold entries of the UtxoEntry type.
